+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 UPDATES:
 
+- [#20553](https://redmine.hwtool.net/issues/20553) Add forwarding headers X-Consul-*
 - [#20559](https://redmine.hwtool.net/issues/20559) Add method "getAll()" to ServiceInstances class
 - [#20521](https://redmine.hwtool.net/issues/20521) Fix bug with default timeout for Consul Watch
 - [#20123](https://redmine.hwtool.net/issues/20123) Add to "consul-service-health-monitor" module functionality of an auto-reconnect to Сonsul